+++ eslint.config.mjs
@@ -1,0 +1,32 @@
+import eslint from "@eslint/js";
+import reactPlugin from "eslint-plugin-react";
+import tseslint from "typescript-eslint";
+
+export default tseslint.config(
+  eslint.configs.recommended,
+  tseslint.configs.recommendedTypeChecked,
+  {
+    languageOptions: {
+      parserOptions: {
+        projectService: true,
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
+  },
+  reactPlugin.configs.flat.recommended,
+  reactPlugin.configs.flat['jsx-runtime'],
+  {
+    settings: {
+      react: {
+        version: "detect",
+      },
+    },
+    rules: {
+      "no-console": "warn",
+      "no-duplicate-imports": "error",
+      "react/destructuring-assignment": "error",
+      "react/jsx-no-useless-fragment": "error",
+      "@typescript-eslint/no-unused-vars": "off",
+    }
+  }
+);