+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+_Fev. 24, 2025_
+
+- setup typescript-eslint and prettier
+- add custom rules
+
 ## 0.1.0
 
 _Jan. 29, 2025_